The environment-specific ii18n configuration are now isolated into the respective environment files
