+++ config/environments/development.rb
@@ -1,4 +1,18 @@
 require "active_support/core_ext/integer/time"
+require_relative "../../app/models/contentful_client"
+require_relative "../../lib/i18n/backend/contentful"
+
+module I18n
+  class CustomExceptionHandler < ExceptionHandler
+    def call(exception, locale, key, options)
+      if exception.is_a?(MissingTranslation) && key.to_s != "i18n.plural.rule"
+        raise exception.to_exception
+      else
+        super
+      end
+    end
+  end
+end
 
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
@@ -45,4 +59,13 @@ Rails.application.configure do
 
   # Apply autocorrection by RuboCop to files generated by `bin/rails generate`.
   # config.generators.apply_rubocop_autocorrect_after_generate!
+
+  config.before_initialize do
+    # Set the I18n backend
+    yaml_backend = I18n::Backend::Simple.new
+    I18n.backend = I18n::Backend::Chain.new(yaml_backend)
+
+    I18n.exception_handler = I18n::CustomExceptionHandler.new
+    Rails.logger.info "Development: Custom before_initialize logic executed for I18n."
+  end
 end
